Extract director specific to tab bar management
